simple_chatbot/pyproject.toml
[project]
name = "simple-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.6.4",
    "openai-agents>=0.2.5",
    "python-dotenv>=1.1.1",
]
